requirements.txt
requests
pwntools
termcolor
